fix(endpoint): integrate BulkDsl into default endpoint.* API, remove mirrors and NestPrefix
