+++ endpoint/shared/src/test/scala-3.7/zio/blocks/endpoint/EndpointGroupSpec.scala
@@ -17,7 +17,6 @@
 package zio.blocks.endpoint
 
 import zio.blocks.endpoint.*
-import zio.blocks.endpoint.BulkDsl.*
 import zio.blocks.endpoint.RoutePattern.*
 import zio.http.Method
 import zio.test.*
